\name{checkNoException}
\alias{checkNoException}
\title{
Checking for occurrence of exceptions
}
\description{
Checks that no exceptions occurred
}
\usage{
checkNoException(expr)
}
\arguments{
  \item{expr}{Expression to be executed}
}
\value{
Returns \code{TRUE} if no exceptions occurred, otherwise throws an exception.
}
\author{
Martin Studer\cr
Mirai Solutions GmbH \url{http://www.mirai-solutions.com}
}
\keyword{utilities}
\keyword{internal}
