\name{mirai}
\alias{mirai}
\docType{data}
\title{
Mirai Solutions GmbH
}
\description{
Utility object to easily get to the Mirai Solutions GmbH web page. Just enter \code{mirai} in the R console.
}
\usage{mirai}
\references{
Mirai Solutions GmbH \url{https://mirai-solutions.ch}
}
\keyword{datasets}
