\name{rsrc}
\alias{rsrc}
\title{
Constructing Excel area references
}
\description{
Constructs an Excel area reference
}
\usage{
rsrc(resource)
}
\arguments{
  \item{resource}{File or directory path relative to unit test resource directory}
}
\value{
Returns full path to unit test resource
}
\author{
Martin Studer\cr
Mirai Solutions GmbH \url{http://www.mirai-solutions.com}
}
\keyword{utilities}
\keyword{internal}
