\name{swissfranc}
\docType{data}
\alias{swissfranc}
\title{Historical Exchange Rates: CHF vs EUR, USD and GBP}
\description{
  This data set provides historical exchange rates (CHF vs EUR, USD, GBP)
  in the time frame from January 1, 2014 to February 24, 2015. The exchange 
  rates reflect bid prices with a 0\% interbank rate.
}
\usage{swissfranc}
\format{A data.frame with daily exchange rates in the mentioned time frame.}
\source{retrieved via \samp{https://www.oanda.com/} - the retrieved time range is no longer available.}
\keyword{datasets}
