\name{XLConnect-deprecated}
\alias{XLConnect-deprecated}
\title{Deprecated functions in package \pkg{XLConnect}}
\description{
These functions are provided for compatibility with older versions of
\pkg{XLConnect} only, and will be defunct in a later release.
}
\details{
The following functions are deprecated and will be made defunct. Use the replacements as indicated.
\itemize{
  \item \code{getReferenceCoordinates}: \code{\link[=getReferenceCoordinatesForName-methods]{getReferenceCoordinatesForName}}
}
}
