\name{XLConnect-package}
\alias{XLConnect-package}
\alias{XLConnect}
\docType{package}
\title{
XLConnect - Reading and writing Microsoft Excel files from R
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab XLConnect\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2010-07-30\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Martin Studer, Mirai Solutions GmbH

Maintainer: Mirai Solutions GmbH <info@mirai-solutions.com>
}
\references{
Literature or other references for background information
}
\keyword{ 
package
file
}
\examples{
search()
}
